Moves only function from handle_vii.py into NowcastLightning plugin along with associated unit tests.
